% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_markers}
\alias{heatmap_markers}
\title{Theme for publication}
\usage{
heatmap_markers(
  markers.matrix = NULL,
  color = c("blue", "white", "red"),
  paletteLength = 100,
  fontsize_row = 10,
  color.rows = T,
  title = "log(FC)",
  add.flag = T,
  de.cutoff = 4,
  top.n = NULL,
  repel.degree = 0,
  legend = T,
  annotation_legend = T,
  cellwidth = NA,
  cellheight = NA,
  save = F,
  width = 7,
  height = 11,
  file = "/home/figo/software/scell_skin_inflammation_master/plot_out/IM11/p.pdf"
)
}
\arguments{
\item{base_size}{default font size is 12}

\item{base_family}{default font is "HelveticaNeueLT Std" that requires to be install in the system}

\item{axis}{TRUE}

\item{grid}{default without grid, is FALSE, or "major", "minor"}

\item{legend.position}{default is "none", could be "top", "right", "bottom", or "left"}
}
\description{
Theme for publication
}
