% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pub_colors}
\alias{pub_colors}
\title{Custom fill color}
\usage{
pub_colors(alpha = 1, n = 1, ...)
}
\arguments{
\item{alpha}{alpha value (0, 1)}

\item{n}{could be vector ranging 1:35}

\item{values}{a list of color values}
}
\description{
Custom fill color
}
