% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pub_pal1}
\alias{pub_pal1}
\title{Get palette colors}
\usage{
pub_pal1(
  limits = c(0, 1),
  aesthetics = c("color", "colour", "fill"),
  palette = c("RdBu", "BrBG", "PiYG", "PRGn", "PuOr", "RdYlBu"),
  n = 200
)
}
\arguments{
\item{limits}{rescale values, default is c(0, 1)}

\item{aesthetics}{c("color", "fill")}

\item{palette}{palette color palettes}

\item{n}{the number of colors (>= 1) to be in the palette.}
}
\value{
A character vector containing color names
}
\description{
Get palette colors from palette theme name and n. The color palettes
are from RColorBrewer, but with the middle color changing to '#FFFFFF'(white),
thus we can visualize element 0 with white color.
palette colors are suitable for visualize a matrix which elements are partly positive and
partly negative (e.g. correlation matrix in \link{-1, 1}, or \link{-20, 100}).
}
\seealso{
Function \code{\link{colorRamppalette}}, package \code{RColorBrewer}
}
\keyword{color}
