% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pub_pal2}
\alias{pub_pal2}
\title{Get palette colors}
\usage{
pub_pal2(
  limits = c(0, 1),
  aesthetics = c("color", "colour", "fill"),
  palette = c("Oranges", "Purples", "Reds", "Blues", "Greens", "Greys", "OrRd", "YlOrRd",
    "YlOrBr", "YlGn"),
  n = 200
)
}
\arguments{
\item{limits}{rescale values, default is c(0, 1)}

\item{aesthetics}{c("color", "fill")}

\item{palette}{palette color palettes}

\item{n}{the number of colors (>= 1) to be in the palette.}
}
\value{
A character vector containing color names
}
\description{
Get palette colors from palette theme name and n. The color palettes
are from RColorBrewer. palette colors are suitable for visualize a non-negative
or non-positive matrix (e.g. matrix in \link{0, 20}, or \link{-100, -10}, or \link{100, 500}).
}
\seealso{
Function \code{\link{colorRamppalette}}, package \code{RColorBrewer}
}
\keyword{color}
