% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pub_scale_colors}
\alias{pub_scale_colors}
\title{Custom scale_fill_manual, scale_colour_manual}
\usage{
pub_scale_colors(alpha = 1, ...)
}
\arguments{
\item{alpha}{alpha value (0, 1)}

\item{values}{a list of color values}
}
\description{
Custom scale_fill_manual, scale_colour_manual
}
