% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{save_pdf}
\alias{save_pdf}
\title{Theme for publication}
\usage{
save_pdf(x, filename, width = 7, height = 7)
}
\arguments{
\item{x}{the last plot}

\item{filename}{the filename}

\item{width}{default is 7}

\item{height}{default is 7}
}
\description{
Theme for publication
}
