% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomThemePublication.R
\name{scale_color_publication}
\alias{scale_color_publication}
\title{Custom discrete color}
\usage{
scale_color_publication(...)
}
\arguments{
\item{values}{a list of color values}
}
\description{
Custom discrete color
}
