% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomThemePublication.R
\name{scale_fill_publication}
\alias{scale_fill_publication}
\title{Custom fill color}
\usage{
scale_fill_publication(...)
}
\arguments{
\item{values}{a list of color values}
}
\description{
Custom fill color
}
