% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split2xls.R
\name{split2xls}
\alias{split2xls}
\title{Split Dataframe into Excel Files and Sheets}
\usage{
split2xls(
  df,
  file_col = NULL,
  sheet_col,
  output_dir = file.path(getwd(), "split2xls/split2xls.xlsx"),
  gene_col = "gene",
  gene_annotation = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{df}{Dataframe to be split.}

\item{file_col}{(Optional) Column used to determine separate Excel files. If NULL, all sheets go into one file.}

\item{sheet_col}{Column used to determine different sheets within each Excel file.}

\item{output_dir}{Directory and file names where Excel files should be saved (default: "split2xls/").}

\item{gene_col}{Name of the SYMBOL gene column for annotation (default: "gene").}

\item{gene_annotation}{Logical. If TRUE, merges additional gene annotation data from uniprot and refseq (default: TRUE).}

\item{overwrite}{Logical. If TRUE, overwrites existing Excel files (default: TRUE).}
}
\value{
Creates Excel files in the specified output directory.
}
\description{
Splits a dataframe into separate Excel files based on one column and organizes sheets within each file based on another column.
}
