% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thepubr.R
\name{theme_border}
\alias{theme_border}
\title{Custom theme}
\usage{
theme_border(
  base_size = 12,
  base_family = "HelveticaNeueLT Std",
  border = TRUE,
  axis = TRUE,
  grid = FALSE,
  legend.position = "top",
  facet_background = "dodgerblue4",
  facet_border = "black",
  facet_color = "white",
  rotate_text = NULL
)
}
\arguments{
\item{base_size}{default font size is 12}

\item{base_family}{default font is : "HelveticaNeueLT Std"}

\item{border}{TRUE: plot has border}

\item{axis}{TRUE}

\item{grid}{default is FALSE, or "major", "minor", "all"}

\item{legend.position}{default is none}

\item{facet_background}{defalut is 'dodgerblue4'}

\item{facet_border}{defalut is 'black'}

\item{facet_color}{default is 'white'}
}
\description{
Custom theme
}
