% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thepubr.R
\name{theme_publication}
\alias{theme_publication}
\title{Theme for publication}
\usage{
theme_publication(
  font.size = 12,
  font.family = "HelveticaNeueLT Std",
  axis = TRUE,
  grid = FALSE,
  legend.position = "top",
  legend.font.size = 12,
  aspect.ratio = NULL,
  rotate.text = NULL
)
}
\arguments{
\item{font.size}{default font size is 12}

\item{font.family}{default font is "HelveticaNeueLT Std" that requires to be install in the system}

\item{axis}{TRUE}

\item{grid}{default without grid, is FALSE, or "major", "minor", "all"}

\item{legend.position}{default is "top", could be "none", "top", "right", "bottom", or "left"}

\item{legend.font.size}{default legend font size is 12}

\item{aspect.ratio}{default is "1": square}
}
\description{
Theme for publication
}
