% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{thepubr_colors}
\alias{thepubr_colors}
\title{Custom fill color}
\usage{
thepubr_colors(alpha = 1, n = 1, ...)
}
\arguments{
\item{alpha}{alpha value (0, 1)}

\item{n}{could be vector ranging c(1, 35)}

\item{values}{a list of color values}
}
\description{
Custom fill color
}
