% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pub_scale_colors}
\alias{pub_scale_colors}
\title{Custom scale_fill_manual, scale_colour_manual}
\usage{
pub_scale_colors(alpha = 1, aes = "colour", ...)
}
\arguments{
\item{alpha}{alpha value (0, 1)}

\item{aes}{default is "colour" with scale_colour_manual,
use aesthetic mapping with "colour" like geom_line() and geom_point().
If aes = "fill" with scale_fill_manual,
use aesthetic mapping with "fill" like geom_bar(), geom_area(), and geom_polygon().}

\item{values}{a list of color values}
}
\description{
Custom scale_fill_manual, scale_colour_manual
}
