% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thepubr.R
\name{save_figure}
\alias{save_figure}
\title{Custom save figure}
\usage{
save_figure(
  plot = last_plot(),
  filename,
  device = "pdf",
  units = "in",
  size = c("custom", "small", "medium", "large", "small_wide", "medium_wide",
    "large_wide", "small_long", "medium_long", "large_long", "super"),
  width = 2.5,
  height = 2.5,
  gg = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{filename}{subfolder/filename}

\item{size}{could be "small", "medium", "large", "super; "small_wide", "small_long"...}

\item{width}{custom width}

\item{height}{custom height}
}
\description{
Custom save figure
}
