% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomThemePublication.R
\name{save_figure}
\alias{save_figure}
\title{Custom save figure}
\usage{
save_figure(
  plot = last_plot(),
  filename,
  device = "pdf",
  units = "in",
  directory = dirname(getwd()),
  size = "small",
  gg = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{filename}{subfolder/filename}

\item{directory}{default is parent directory}

\item{size}{default is "small", could be "small", "medium", "large", _ "wide", "long"}
}
\description{
Custom save figure
}
