% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thepubr.R
\name{scale_color_publication}
\alias{scale_color_publication}
\title{Custom fill color}
\usage{
scale_color_publication(alpha = 1, ...)
}
\arguments{
\item{values}{a list of color values}
}
\description{
Custom fill color
}
