% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thepubr.R
\name{theme_border}
\alias{theme_border}
\title{Custom theme}
\usage{
theme_border(
  font.size = 12,
  font.family = "HelveticaNeueLT Std",
  border = TRUE,
  axis = TRUE,
  grid = FALSE,
  legend.position = "top",
  legend.font.size = 12,
  facet.background = "#F0F0F0",
  facet.border = "black",
  facet.color = "white",
  aspect.ratio = NULL,
  rotate.text = NULL
)
}
\arguments{
\item{font.size}{default font size is 12}

\item{font.family}{default font is : "HelveticaNeueLT Std"
#' @param rotate.text (optional) numeric angle to rotate x-axis labels; default NULL}

\item{border}{TRUE: plot has border}

\item{axis}{TRUE}

\item{grid}{default is FALSE, or "major", "minor", "all"}

\item{legend.position}{default is "top", could be "none", "top", "right", "bottom", or "left"}

\item{legend.font.size}{default legend font size is 12}

\item{facet.background}{default is 'grey85'}

\item{facet.border}{default is 'black'}

\item{facet.color}{default is 'white'}

\item{rotate.text}{(optional) numeric angle to rotate x-axis labels; default NULL}
}
\description{
Custom theme
}
