% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomThemePublication.R
\name{theme_publication}
\alias{theme_publication}
\title{Theme for publication}
\usage{
theme_publication(
  base_size = 12,
  base_family = "Helvetica",
  axis = TRUE,
  grid = FALSE,
  legend.position = "none",
  rotate_text = "none"
)
}
\arguments{
\item{base_size}{default font size is 12}

\item{base_family}{default font is "Arial}

\item{grid}{plot has grid}

\item{legend.position}{default is "none", could be "top", "right", "bottom", or "left"}
}
\description{
Theme for publication
}
