% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thepubr.R
\name{theme_publication}
\alias{theme_publication}
\title{Theme for publication}
\usage{
theme_publication(
  base_size = 12,
  base_family = "HelveticaNeueLT Std",
  axis = TRUE,
  grid = FALSE,
  legend.position = "top",
  rotate_text = NULL
)
}
\arguments{
\item{base_size}{default font size is 12}

\item{base_family}{default font is "HelveticaNeueLT Std" that requires to be install in the system}

\item{axis}{TRUE}

\item{grid}{default without grid, is FALSE, or "major", "minor"}

\item{legend.position}{default is "none", could be "top", "right", "bottom", or "left"}
}
\description{
Theme for publication
}
